requirements.txt
gradio==4.44.0
torch==2.1.0
transformers==4.36.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.1.0
langchain-community==0.0.10
numpy==1.24.3